+++ docs/content/en/integrations/parsers/file/Jfrog_xray_on_demand_binary Scan.md
@@ -2,4 +2,8 @@
 title: "Jfrog xray on demand binary Scan"
 toc_hide: true
 ---
-Import the JSON format for the "\Jfrog_xray_on_demand_binary Scan\" file. Use this importer for Xray version 2.X
+Import the JSON format for the "\Jfrog xray on demand binary Scan\" file. Use this importer for Xray version 2.X
+-- 
+ jfron file documentation
+
+https://jfrog.com/help/r/jfrog-cli/downloading-updates-for-xray-s-database
