+++ dojo/settings/settings.dist.py
@@ -1344,7 +1344,7 @@ HASHCODE_FIELDS_PER_SCANNER = {
     'Bugcrowd API Import': ['unique_id_from_tool'],
     'Rubocop Scan': ['vuln_id_from_tool', 'file_path', 'line'],
     'JFrog Xray Scan': ['title', 'description', 'component_name', 'component_version'],
-    'JFrog Xray': ['title', 'description', 'component_name', 'component_version'],
+    'Jfrog Xray On Demand Binary Scan': ['title', 'description', 'component_name', 'component_version'],
     'CycloneDX Scan': ['vuln_id_from_tool', 'component_name', 'component_version'],
     'SSLyze Scan (JSON)': ['title', 'description'],
     'Harbor Vulnerability Scan': ['title'],
@@ -1543,7 +1543,7 @@ DEDUPLICATION_ALGORITHM_PER_PARSER = {
     'Bugcrowd API': DEDUPE_ALGO_UNIQUE_ID_FROM_TOOL,
     'Rubocop Scan': DEDUPE_ALGO_HASH_CODE,
     'JFrog Xray Scan': DEDUPE_ALGO_HASH_CODE,
-    'JFrog Xray Vulnerabilities': DEDUPE_ALGO_HASH_CODE,
+    'Jfrog Xray On Demand Binary Scan': DEDUPE_ALGO_HASH_CODE,
     'CycloneDX Scan': DEDUPE_ALGO_HASH_CODE,
     'SSLyze Scan (JSON)': DEDUPE_ALGO_HASH_CODE,
     'Harbor Vulnerability Scan': DEDUPE_ALGO_HASH_CODE,
