--- unittests/scans/jfrog_xray_vulnerabilities/many_vulns.json
@@ -1,122 +1,0 @@
-[
-    {
-      "scan_id": "bf080b35-dd8f-4927-5db6-fb188ae8d984",
-      "vulnerabilities": [
-        {
-          "cves": [
-            {
-              "cve": "CVE-2017-18640",
-              "cvss_v2_score": "5.0",
-              "cvss_v2_vector": "CVSS:2.0/AV:N/AC:L/Au:N/C:N/I:N/A:P",
-              "cvss_v3_score": "7.5",
-              "cvss_v3_vector": "CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:H"
-            }
-          ],
-          "summary": "The Alias feature in SnakeYAML 1.18 allows entity expansion during a load operation, a related issue to CVE-2003-1564.",
-          "severity": "High",
-          "components": {
-            "gav://org.yaml:snakeyaml:1.16": {
-              "fixed_versions": [
-                "[1.26]"
-              ],
-              "impact_paths": [
-                [
-                  {
-                    "component_id": "gav://co.com.bancolombia.hygieia:core:1.0.0-SNAPSHOT"
-                  },
-                  {
-                    "component_id": "gav://org.yaml:snakeyaml:1.16",
-                    "full_path": "lib/snakeyaml-1.16.jar"
-                  }
-                ]
-              ]
-            }
-          },
-          "issue_id": "XRAY-92904",
-          "references": [
-            "https://lists.fedoraproject.org/archives/list/package-announce@lists.fedoraproject.org/message/PTVJC54XGX26UJVVYCXZ7D25X3R5T2G6/",
-            "https://lists.fedoraproject.org/archives/list/package-announce@lists.fedoraproject.org/message/CKN7VGIKTYBCAKYBRG55QHXAY5UDZ7HA/",
-            "https://bitbucket.org/asomov/snakeyaml/issues/377/allow-configuration-for-preventing-billion",
-            "https://bitbucket.org/asomov/snakeyaml/wiki/Billion%20laughs%20attack",
-            "https://mvnrepository.com/artifact/org.yaml/snakeyaml/1.25/usages"
-          ]
-        },
-        {
-          "cves": [
-            {
-              "cve": "CVE-2014-0114",
-              "cvss_v2_score": "7.5",
-              "cvss_v2_vector": "CVSS:2.0/AV:N/AC:L/Au:N/C:P/I:P/A:P"
-            }
-          ],
-          "summary": "Apache Commons BeanUtils, as distributed in lib/commons-beanutils-1.8.0.jar in Apache Struts 1.x through 1.3.10 and in other products requiring commons-beanutils through 1.9.2, does not suppress the class property, which allows remote attackers to \"manipulate\" the ClassLoader and execute arbitrary code via the class parameter, as demonstrated by the passing of this parameter to the getClass method of the ActionForm object in Struts 1.",
-          "severity": "High",
-          "components": {
-            "gav://commons-beanutils:commons-beanutils:1.9.2": {
-              "fixed_versions": [
-                "[1.9.4]"
-              ],
-              "impact_paths": [
-                [
-                  {
-                    "component_id": "gav://co.com.bancolombia.hygieia:core:1.0.0-SNAPSHOT"
-                  },
-                  {
-                    "component_id": "gav://commons-beanutils:commons-beanutils:1.9.2",
-                    "full_path": "lib/commons-beanutils-1.9.2.jar"
-                  }
-                ]
-              ]
-            }
-          },
-          "issue_id": "XRAY-55616",
-          "references": [
-            "http://advisories.mageia.org/MGASA-2014-0219.html",
-            "https://issues.apache.org/jira/browse/BEANUTILS-463",
-            "https://security.gentoo.org/glsa/201607-09"
-          ]
-        },
-        {
-          "cves": [
-            {
-              "cvss_v2_score": "7.5",
-              "cvss_v2_vector": "CVSS:2.0/AV:N/AC:L/Au:N/C:P/I:P/A:P"
-            }
-          ],
-          "summary": "Spring Boot ErrorMvcAutoConfiguration.java whitelabel Error Page SpEL Injection ",
-          "severity": "High",
-          "components": {
-            "gav://org.springframework.boot:spring-boot-autoconfigure:1.3.0.RELEASE": {
-              "fixed_versions": [
-                "[1.2.8.RELEASE]",
-                "[1.3.1.RELEASE]"
-              ],
-              "impact_paths": [
-                [
-                  {
-                    "component_id": "gav://co.com.bancolombia.hygieia:core:1.0.0-SNAPSHOT"
-                  },
-                  {
-                    "component_id": "gav://org.springframework.boot:spring-boot-autoconfigure:1.3.0.RELEASE",
-                    "full_path": "lib/spring-boot-autoconfigure-1.3.0.RELEASE.jar"
-                  }
-                ]
-              ]
-            }
-          },
-          "issue_id": "XRAY-79870",
-          "references": [
-            "http://spring.io/blog/2015/12/18/spring-boot-1-3-1-and-1-2-8-available-now",
-            "https://github.com/spring-projects/spring-boot/issues/4763",
-            "https://github.com/spring-projects/spring-boot/commit/edb16a13ee33e62b046730a47843cb5dc92054e6",
-            "https://issues.apache.org/jira/browse/KYLIN-1893",
-            "https://issues.apache.org/jira/browse/KYLIN-1867"
-          ]
-        }
-      ],
-      "component_id": "gav://co.com.bancolombia.hygieia:core:1.0.0-SNAPSHOT",
-      "package_type": "Maven",
-      "status": "completed"
-    }
-  ]
-
