--- unittests/scans/jfrog_xray_vulnerabilities/one_vuln.json
@@ -1,50 +1,0 @@
-[
-    {
-      "scan_id": "bf080b35-dd8f-4927-5db6-fb188ae8d984",
-      "vulnerabilities": [
-        {
-          "cves": [
-            {
-              "cve": "CVE-2017-18640",
-              "cvss_v2_score": "5.0",
-              "cvss_v2_vector": "CVSS:2.0/AV:N/AC:L/Au:N/C:N/I:N/A:P",
-              "cvss_v3_score": "7.5",
-              "cvss_v3_vector": "CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:N/I:N/A:H"
-            }
-          ],
-          "summary": "The Alias feature in SnakeYAML 1.18 allows entity expansion during a load operation, a related issue to CVE-2003-1564.",
-          "severity": "High",
-          "components": {
-            "gav://org.yaml:snakeyaml:1.16": {
-              "fixed_versions": [
-                "[1.26]"
-              ],
-              "impact_paths": [
-                [
-                  {
-                    "component_id": "gav://co.com.bancolombia.hygieia:core:1.0.0-SNAPSHOT"
-                  },
-                  {
-                    "component_id": "gav://org.yaml:snakeyaml:1.16",
-                    "full_path": "lib/snakeyaml-1.16.jar"
-                  }
-                ]
-              ]
-            }
-          },
-          "issue_id": "XRAY-92904",
-          "references": [
-            "https://lists.fedoraproject.org/archives/list/package-announce@lists.fedoraproject.org/message/PTVJC54XGX26UJVVYCXZ7D25X3R5T2G6/",
-            "https://lists.fedoraproject.org/archives/list/package-announce@lists.fedoraproject.org/message/CKN7VGIKTYBCAKYBRG55QHXAY5UDZ7HA/",
-            "https://bitbucket.org/asomov/snakeyaml/issues/377/allow-configuration-for-preventing-billion",
-            "https://bitbucket.org/asomov/snakeyaml/wiki/Billion%20laughs%20attack",
-            "https://mvnrepository.com/artifact/org.yaml/snakeyaml/1.25/usages"
-          ]
-        }
-      ],
-      "component_id": "gav://co.com.bancolombia.hygieia:core:1.0.0-SNAPSHOT",
-      "package_type": "Maven",
-      "status": "completed"
-    }
-  ]
-
